chore: wip

+++ barretenberg/cpp/src/barretenberg/dsl/acir_format/sha256_constraint.test.cpp
@@ -18,12 +18,12 @@ TEST_F(Sha256Tests, TestSha256Compression)
 {
 
     std::array<Sha256Input, 16> inputs;
-    for (uint32_t i = 0; i < 16; ++i) {
-        inputs[i] = { .witness = (i + 1), .num_bits = 32 };
+    for (size_t i = 0; i < 16; ++i) {
+        inputs[i] = { .witness = static_cast<uint32_t>(i + 1), .num_bits = 32 };
     }
     std::array<Sha256Input, 8> hash_values;
-    for (uint32_t i = 0; i < 8; ++i) {
-        hash_values[i] = { .witness = (i + 17), .num_bits = 32 };
+    for (size_t i = 0; i < 8; ++i) {
+        hash_values[i] = { .witness = static_cast<uint32_t>(i + 17), .num_bits = 32 };
     }
     Sha256Compression sha256_compression{
         .inputs = inputs,
